lib/udelay.c: a serious generic udelay
This udelay uses initial calibration, as an initcall. to calculate a
"loops per jiffy" variable.  Calibration takes around 40 jiffies
at startup and is only performed if udelay is actually used.

As of this commit arch-lpc2104 still uses its own arch-specific udelay,
which is kept to actually test the code for arch-specific udelay.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>