general (Warning: breaks flash builds): build bathos.o first
This commit builds bathos.o with a generic lds for all archs.
The choice allows factorizing code about special sections (initcalls and
tasks) leaving only arch-specific addresses in the arch-specific
linker script.

This paves the way for arch-unix, running bathos as a process.

It breaks flash builds because rom initcalls are not part of the
initcall stanza. The next commit fixes the issue, and I prefer
to keep it separate, for documenting purposes.

When I initially wrote the code I had no flash builds, so the same
initcall sequence was used for everyone (and still is, as of this
commit).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>