arch-atmega: add uart-mux-yun device
On the arduino yun board, the atmega32u4 UART port is connected to
the first UART of the ath331 MIPS CPU, used both as Linux console port
and as an mcuio bus. This device lets the user switch between "console"
mode and "mcuio" mode. In console mode, characters coming from the atheros
UART are redirected to bathos_stdout (a pipe attached to the usb-uart
device). On the other hand, in mcuio mode, data coming from the atheros can
be read from the device via pipe_read() (task-mcuio uses this method to
receive input data). An mcuio_data_ready event is triggered when data can
be read from the device.
When working in console mode, characters coming from bathos_stdin (an
usb-uart based pipe again) are redirected to the mips console.