add scripts for automatic generation of obj file containing gpios labels
This script generates a .o file given a gpio cfg file with the following
format:

* Comment lines permitted (shall begin with an hash character).
* One line per gpio.
* Format of a line is: "NAME",capabilities,events

where NAME is a string of exactly 4 chars and capabilities is a string with
the following format

DISABLED |
CAPABILITY_LIST

and CAPABILITY_LIST is:

CAP |
CAP|CAPABILITY_LIST

where CAP is one of the following:

INPUT, OUTPUT, PULLUP, PULLDOWN, ODRAIN

INPUT means that the gpio can be configured as an input.
OUTPUT means that the gpio can be configured as an output.
PULLUP means that a pullup can be activated on this gpio.
PULLDOWN means that a pulldown cna be activated on this gpio
ODRAIN means that this gpio can be configured as open drain

events is a string with the following format

NOEVT|
EVENTS_LIST

and EVENTS_LIST is

EVT|
EVT|EVENTS_LIST

where EVT is one of the following:

RISEDGE, FALLEDGE

RISEDGE means rising edge
FALLEDGE means falling edge