udelay: make init possible even if interrupts are needed for jiffies.
If jiffies are incremented under the periodic timer interrupt,
generic_udelay_init() hangs because it gets called when interrupts are
still disabled (it is an initcall). Implement an udelay_init() which shall
be called with interrupts enabled (and jiffies running) if
ARCH_NEEDS_INTERRUPTS_FOR_JIFFIES is defined.

This is a fast but incorrect solution to the problem. We should use a hw
timer for udelay calibration.