lib/event: improve interrupt events management
Before this commit, a low level handler was mandatory, because otherwise the
high level handler would never be reached (the irq line would get stuck
and the cpu would just go on spinning in and out of the low level handler).
This commit automatically masks the irq line after calling the low level
handler and unmasks it after the high level handler has returned.
Furthermore, the interrupt flags are not copied to a temporary location
anymore, because there's no possible race now (each irq line is disabled
before the relevant interrupt event is handled).
Code and comments are also improved a little bit.