EVENTS: add hook for events notification to main loop
On the esp8266, bathos runs as a task of the main operating system, and
has no infinite main loop. We need to wake up the bathos task each time
an event takes place, so we add this hook which is unused by default and
makes it possibile to do what we have to do on the esp8266 without breaking
other archs (hopefully).