MAIN: separate main loop from bathos_main()
This patch extracts the core of the main loop and implements it as
a weak symbol, so that architectures can override it.
This is needed on the esp8266 because on such architecture the main loop
cannot exist. We just have an esp8266 task which is woken up whenever an
event has to be processed.